see prithivi_now saving issue is solved go in tripwindowdelivery class , i didnot make cjanges in other algos

+++ tripplanning-ms/tripplanning-lib/src/main/java/com/flashex/tripplanningmicroservice/lib/model/Customer.java
@@ -31,7 +31,7 @@ public class Customer {
     public String emailId;
 
     /** Integer variables **/
-    @CassandraType(type = DataType.Name.INT)
-    public int phoneNumber;
+    @CassandraType(type = DataType.Name.BIGINT)
+    public long phoneNumber;
 
 }
